Initial States

MajoranaPropagation.FockState — Type

FockState

A struct to represent a Fock basis state.

MajoranaPropagation.FockState — Method

FockState(n_sites::Int, up_occupied_sites_iter, down_occupied_sites_iter)

Create a spinful Fock basis state given an iterable with the occupied sites for spin-up and a list of occupied sites for spin-down fermions.

MajoranaPropagation.FockState — Method

FockState(n_sites::Int, occupied_sites_iter)

Create a spinless Fock basis state given an iterable with the occupied sites.

MajoranaPropagation.FockState — Method

FockState(n_sites::Integer, symb::Symbol, is_spinful::Bool; hole_positions=nothing, kwargs...)

Create a Fock basis state from the named filling pattern symb. The only supported symbol so far is :checkerboard, an antiferromagnetic pattern: for spinful systems, alternating spin-up and spin-down fermions on a grid with nx sites per row (nx = n_sites if not given); for spinless systems, every odd site occupied. Sites listed in hole_positions are left empty.

MajoranaPropagation.fock_mask — Method

fock_mask(msum::MajoranaSum)

Given a MajoranaSum, return a new MajoranaSum with terms that have an overlap with Fock states.

MajoranaPropagation.overlapwithfock — Method

overlapwithfock(msum::AbstractMajoranaSum, fock_state::FockState)

Compute the expectation value $\langle F | msum | F \rangle$ of the Majorana sum in the Fock basis state $|F\rangle$ given by fock_state.

MajoranaPropagation.overlapwithfock — Method

overlapwithfock(msum::MajoranaSum, sites_with_particle_superposition::Vector{FockState}, superposition_coefficients::Vector{<:Union{Real,Complex}})

Compute the overlap <superposition|msum|superposition> where

superposition is given as a vector of Fock basis states sites_with_particle_superposition

superposition_coefficients are the coefficients of the superposition (assumed normalized)

MajoranaPropagation.overlapwithfock — Method

overlapwithfock(msum::MajoranaSum{TT,CT}, fock_state_1::FockState, fock_state_2::FockState) where {TT<:Integer,CT}

Evaluate the matrix element $\langle F_1 | msum | F_2 \rangle$ of the Majorana sum between the two Fock basis states given by fock_state_1 and fock_state_2.

MajoranaPropagation.overlapwithfock — Method

overlapwithfock(ms::TT, fock_state_1::FockState, fock_state_2::FockState, n_fermions) where {TT<:Integer}

Evaluate the matrix element $\langle F_1 | ms | F_2 \rangle$ of a single Majorana string between the two Fock basis states given by fock_state_1 and fock_state_2.

MajoranaPropagation.overlapwithfock — Method

overlapwithfock(ms::TT, unpaired_mask::TT, fock_state::FockState) where {TT<:Integer}

Compute the expectation value $\langle F | ms | F \rangle$ of a single Majorana string in the Fock basis state $|F\rangle$ given by fock_state. unpaired_mask is the mask created by create_unpaired_mask.

MajoranaPropagation.overlapwithfock — Method

overlapwithfock(ms::TT, sites_with_particle_superposition::Vector{<:FockState}, superposition_coefficients::Vector{<:Union{Real,Complex}}, n_fermions, unpaired_mask::TT) where {TT<:Integer}

Compute the overlap $\langle \psi | ms | \psi \rangle$ of a Majorana string with the superposition $|\psi\rangle = \sum_k c_k |F_k\rangle$ of the Fock basis states sites_with_particle_superposition with coefficients superposition_coefficients (assumed normalized). n_fermions is the number of fermions in the system, used for computing the matrix elements of ms between two different Fock states. unpaired_mask is the mask created by create_unpaired_mask, used to detect strings with vanishing diagonal matrix elements.